+++ src/cadqueryhelper/shape/__init__.py
@@ -25,5 +25,6 @@ from .backdrop import backdrop
 from .jersey_barrier import jersey_barrier
 from .vase import vase
 from .teardrop import teardrop
+from .step_pyramid import step_pyramid
 
 from .ring import ring