+++ changes.md
@@ -2,6 +2,9 @@
 
 ## main wip
 
+## 1.7.1
+* Added shape step_pyramid
+
 ## 1.7.0
 * Upgraded to cadquery version 2.6.1
 * Added shape teardrop